requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
colorlog==6.7.0
frozenlist==1.3.3
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
oauthlib==3.2.2
protobuf==4.24.1
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.16
yarl==1.9.2
